+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Cleaned up imports no longer being used
 - Fixed bug:
   - In `ptv_timetable.types.Run.load()`, a `None` value in the raw response of the `interchange` dict was not handled and caused a `TypeError`
+- Updated documentation
 
 ## 0.5.0 (2026-06-01)
 - Breaking change: Python minimum version requirement updated from 3.12 to 3.14