+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.5.1 (2026-__-__)
+- Edited `ptv_timetable.types.ParserError` error message
 - `ptv_timetable.types.RunInterchange.destination_name` may return `None`
 - Cleaned up imports no longer being used
 - Fixed bug: